+++ server/index.ts
@@ -1,11 +1,14 @@
 import express from "express";
 import { ApolloServer } from "apollo-server-express";
 import { buildSchema } from "type-graphql";
-// import Redis from "ioredis";
-// import session from "express-session";
+import redis from "redis";
+import session from "express-session";
+import connectRedis from "connect-redis";
 import cors from "cors";
 import { createConnection } from "typeorm";
 import { UserResolver } from "./resolvers/UserResolver";
+import "dotenv-safe/config";
+
 import {
   Post,
   User,
@@ -20,7 +23,7 @@ const main = async () => {
   const conn = await createConnection({
     type: "postgres",
     url:
-      "postgresql://ulnmey4j2kgpe9k45mtz:MGlziUJpsaLDAGYNCnb3@brbpvg72wcxztaxjapif-postgresql.services.clever-cloud.com:5432/brbpvg72wcxztaxjapif",
+      process.env.DATABASE_URL,
     synchronize: true,
     logging: true,
     entities: [Post, User, Location, Rating, Poll, Comment, Follower],
@@ -28,12 +31,31 @@ const main = async () => {
 
   const app = express();
 
+  // const RedisStore = connectRedis(session);
+  // let redisClient = redis.createClient()
   app.use(
-    cors({
-      origin: "http://localhost:3000/",
-      credentials: true,
+    session({
+      name: process.env.COOKIE_NAME,
+      // store: new RedisStore({
+      //   client: redisClient,
+      //   disableTouch: true,
+      // }),
+      cookie: {
+        maxAge: 1000 * 60 * 60 * 24 * 365 * 10, // 10 years
+        httpOnly: true,
+        sameSite: "lax", // csrf
+        secure: false, // cookie only works in https
+      },
+      saveUninitialized: false,
+      secret: process.env.SESSION_SECRET || "hi",
+      resave: false,
     })
+  //   cors({
+  //     origin: "http://localhost:3000/",
+  //     credentials: true,
+  //   })
   );
+  
 
   const apolloServer = new ApolloServer({
     schema: await buildSchema({
@@ -43,7 +65,7 @@ const main = async () => {
     context: ({ req, res }) => ({
       req,
       res,
-      // redis,
+      redis,
     }),
   });
