+++ server/types.ts
@@ -1,9 +1,10 @@
 
 import { Request, Response, } from "express";
-// import { createUserLoader } from "./utils/createUserLoader";
+import { Redis } from "ioredis";
 
 export type MyContext = {
     req: Request ;
     res: Response;
+    redis: Redis;
     // userLoader: ReturnType<typeof createUserLoader>;
 };
